feat: ignore privilege error when running with secure boot (#50)
When running with secure boot the iopl() call is blocked by the
kernel. We're already quite sure at that point that we're running
in a VMware VM, so we'll ignore the error in secure boot mode,
assuming the operator knows what they're doing.

This commit includes the results of `make rekres`.

Signed-off-by: Robin Elfrink <robin.elfrink@eu.equinix.com>